pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-agent"
version = "0.1.0"
description = "CLI-first documentation QA agent scaffold"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-community>=0.3.0",
  "langchain-chroma>=0.2.0",
  "langchain-text-splitters>=0.3.0",
  "chromadb>=0.5.0",
]

[project.scripts]
doc-agent = "app.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
